fix(conda): remove Python version tag
Prior to this commit, the `{{ python }}` tag was used in `base-node`,
`base-node-rpc`, and `teensy-minimal-rpc` to dynamically specify the
required Python based on the version of Python used during the build.
However, since configuring these packages to build as `noarch: python`,
the Python version **MUST** not be specified or the package will be locked
to that version.

In this commit, update these modules to remove the Python version tag,
i.e., `{{ python }}` from the build recipe requirements.