fix(nadamq): do not use Conda variants
Prior to this commit, the Appveyor build configuration ran in a Python 3.6
environment and used `variants.yaml` to build for both Python 2.7 and
Python 3.6 in a single job.  However, this seems to result in an incorrect
`vc` package version for the Python 2.7 release, i.e., `vc==14` when it
should be `vc==9` for Python 2.7.

In this commit, do not use Conda variants and add additional Python 2.7
configurations to the Appveyor build matrix entries.